+++ src/main/java/com/lilangel/teamplay/service/UserService.java
@@ -14,7 +14,7 @@ public interface UserService {
      * @param isAdmin    является ли пользователь админом
      * @return ответ с идентификатором пользователя
      */
-    Integer create(Integer tgId, Integer employerId, Boolean isAdmin);
+    Integer create(Long tgId, Integer employerId, Boolean isAdmin);
 
     /**
      * Удаляет пользователя по идентификатору
